Tweak in how the animation is triggered
In order to avoid confusion, I removed the little presentational trick to trigger the animation on hover and switched to a class-based JavaScript-fired trigger. The animation now starts when the call-progress--running class is applied to the div.class-progress element. This action is toggled by clicking on the `<small>` description. There's also minor code reorganization for better clarity.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>